feat: implement item movement and tile change notifications
- Added `get_tile` function to retrieve items at a specific position on the map.
- Enhanced `SessionActor` to handle item movement commands, including validation for adjacent positions and item flags.
- Implemented broadcasting of item movement results and tile changes to players.
- Updated `GameMap` with methods for item management, including `remove_item` and `drop_item`.
- Introduced new item flags and attributes to support container functionality and tile friction.
- Modified player and agent structures to accommodate changes in speed and level types.
- Adjusted serialization of player descriptions to include speed.
- Enhanced position handling with methods to check for container and inventory coordinates.

+++ src/entities/agent.rs
@@ -19,7 +19,7 @@ new_key_type! { pub struct AgentKey; }
 pub struct Agent {
     pub handle: Option<AgentKey>,
     pub next_walk_tick: Tick,
-    pub speed: u32,
+    pub speed: u16,
     inner: AgentInner,
 }
 
@@ -41,7 +41,7 @@ impl Agent {
         }
     }
 
-    pub fn calculate_walk_ticks(&self, tile_friction: u8, diagonal: bool) -> Tick {
+    pub fn calculate_walk_ticks(&self, tile_friction: u32, diagonal: bool) -> Tick {
         let move_speed = (SPEED_PARAM_A * ((self.speed as f32) + SPEED_PARAM_B).ln()
             + SPEED_PARAM_C)
             .round()
